# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Net Zone CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change current user password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed

  # ── CUSTOMERS ───────────────────────────────────────────────────────────────
...
  /calls:
    get:
      operationId: listCalls
      tags: [calls]
      summary: List call records
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, yesterday, week]
        - name: agentId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of calls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Call"
    post:
      operationId: createCall
      tags: [calls]
      summary: Create a call record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallInput"
      responses:
        "201":
          description: Call created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
  /calls/bulk-sync:
    post:
      operationId: bulkSyncCalls
      tags: [calls]
      summary: Bulk sync offline call records
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkSyncInput"
      responses:
        "200":
          description: Synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkSyncResult"
  /calls/{id}:
    put:
      operationId: updateCall
      tags: [calls]
      summary: Update a call record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallUpdate"
      responses:
        "200":
          description: Call updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
    delete:
      operationId: deleteCall
      tags: [calls]
      summary: Delete a call record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── REMARKS ─────────────────────────────────────────────────────────────────
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── DASHBOARD ────────────────────────────────────────────────────────────────
...
  /reports/calls:
    get:
      operationId: getCallsReport
      tags: [reports]
      summary: Calls report with date and agent filters
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Calls report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Call"
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents/users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new agent (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    put:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Deactivate agent (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deactivated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, name, role, isActive]
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, agent]
        isActive:
          type: boolean
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string

    # ── CUSTOMERS ─────────────────────────────────────────────────────────────
...
    Call:
      type: object
      required: [id, customerId, agentId, type, duration, createdAt]
      properties:
        id:
          type: string
        customerId:
          type: string
        customerName:
          type: ["string", "null"]
        agentId:
          type: string
        agentName:
          type: ["string", "null"]
        type:
          type: string
          enum: [incoming, outgoing, missed]
        duration:
          type: integer
        phoneNumber:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        remark:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        reminderDate:
          type: ["string", "null"]
        deviceId:
          type: ["string", "null"]
        createdAt:
          type: string
    CallInput:
      type: object
      required: [customerId, type, duration]
      properties:
        customerId:
          type: string
        type:
          type: string
          enum: [incoming, outgoing, missed]
        duration:
          type: integer
        phoneNumber:
          type: string
        category:
          type: string
        remark:
          type: string
        followUpDate:
          type: string
        reminderDate:
          type: string
        deviceId:
          type: string
    CallUpdate:
      type: object
      properties:
        type:
          type: string
        duration:
          type: integer
        remark:
          type: string
        category:
          type: string
        followUpDate:
          type: string
        reminderDate:
          type: string
    BulkSyncInput:
      type: object
      required: [calls]
      properties:
        calls:
          type: array
          items:
            $ref: "#/components/schemas/CallInput"
    BulkSyncResult:
      type: object
      required: [synced, failed]
      properties:
        synced:
          type: integer
        failed:
          type: integer

    # ── REMARKS ───────────────────────────────────────────────────────────────
...
    Category:
      type: object
      required: [id, name, color, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
        color:
          type: string
        description:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        description:
          type: string

    # ── AGENTS ────────────────────────────────────────────────────────────────
    Agent:
      type: object
      required: [id, username, name, role, isActive, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, agent]
        isActive:
          type: boolean
        totalCalls:
          type: ["integer", "null"]
        todayCalls:
          type: ["integer", "null"]
        createdAt:
          type: string
    AgentInput:
      type: object
      required: [username, name, password, role]
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, agent]
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, agent]
        isActive:
          type: boolean
        password:
          type: string

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...