# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Net Zone CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentUpdateRole } from './agentUpdateRole';

export interface AgentUpdate {
  name?: string;
  role?: AgentUpdateRole;
  isActive?: boolean;
  password?: string;
}